refactor: remove classproperty dependency and improve type safety
