+++ instructor/utils/__init__.py
@@ -1,0 +1,52 @@
+from instructor.utils.gemini_utils import (
+    convert_to_genai_messages,
+    map_to_gemini_function_schema,
+    transform_to_gemini_prompt,
+    update_gemini_kwargs,
+)
+from instructor.utils.json_extract import (
+    extract_json_from_codeblock,
+    extract_json_from_stream,
+    extract_json_from_stream_async,
+)
+from instructor.utils.messages import (
+    dump_message,
+    get_message_content,
+    is_async,
+    merge_consecutive_messages,
+)
+from instructor.utils.provider import Provider, get_provider
+from instructor.utils.system import (
+    SystemMessage,
+    combine_system_messages,
+    extract_genai_system_message,
+    extract_system_messages,
+)
+from instructor.utils.usage import update_total_usage
+
+__all__ = [
+    # Provider
+    "Provider",
+    "get_provider",
+    # JSON
+    "extract_json_from_codeblock",
+    "extract_json_from_stream",
+    "extract_json_from_stream_async",
+    # Messages
+    "dump_message",
+    "get_message_content",
+    "is_async",
+    "merge_consecutive_messages",
+    # System
+    "SystemMessage",
+    "combine_system_messages",
+    "extract_genai_system_message",
+    "extract_system_messages",
+    # Usage
+    "update_total_usage",
+    # Gemini
+    "convert_to_genai_messages",
+    "map_to_gemini_function_schema",
+    "transform_to_gemini_prompt",
+    "update_gemini_kwargs",
+]
