+++ pyproject.toml
@@ -16,7 +16,7 @@ dependencies = [
     "requests<3.0.0,>=2.32.3",
 ]
 name = "instructor"
-version = "1.7.5"
+version = "1.7.8"
 description = "structured outputs for llm"
 readme = "README.md"
